requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
asgiref==3.11.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comtypes==1.2.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
enum34==1.1.10
flatbuffers==25.2.10
fonttools==4.58.0
future==1.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.48.2
gunicorn==25.1.0
httplib2==0.22.0
idna==3.10
jax==0.6.1
jaxlib==0.6.1
kiwisolver==1.4.8
matplotlib==3.10.3
mediapipe==0.10.9
ml_dtypes==0.5.1
MouseInfo==0.1.3
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.9.0.80
opt_einsum==3.4.0
packaging==25.0
pillow==10.2.0
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20181226
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyGetWindow==0.0.9
PyJWT==2.11.0
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pywin32==310
requests==2.32.3
rsa==4.9.1
scipy==1.15.3
screen-brightness-control==0.22.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.2
SpeechRecognition==3.10.1
sqlparse==0.5.5
srt==3.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.3
uritemplate==4.1.1
urllib3==2.4.0
vosk==0.3.45
websockets==15.0.1
WMI==1.5.1
